feat(utils): getGroupedObservable
