+++ packages/utils/src/collect.test.ts
@@ -92,53 +92,4 @@ describe("collect", () => {
       })
     })
   })
-  describe("collect/get", () => {
-    it("returns the inner observable for the giving key", () => {
-      scheduler().run(({ expectObservable, cold }) => {
-        const toGrouped = (source: string, key: string) => {
-          const result = cold(source).pipe(shareReplay(1)) as GroupedObservable<
-            string,
-            string
-          >
-          result.key = key
-          return result
-        }
-
-        const a = toGrouped("-------|     ", "a")
-        const b = toGrouped("---|        ", "b")
-        const c = toGrouped("------| ", "c")
-        const d = toGrouped("------| ", "d")
-
-        const sourceStr = "  (abc)d---------------------|   "
-        const source = cold(sourceStr, { a, b, c, d }).pipe(
-          skip(1),
-          startWith(a),
-        )
-        const result = collect<string, string>()(source)
-        result.subscribe()
-        expectObservable(result.get("a")).toBe("-------|")
-        expectObservable(result.get("b")).toBe("---|")
-        expectObservable(result.get("c")).toBe("------|")
-        expectObservable(result.get("d")).toBe("-----------|")
-      })
-    })
-
-    it("errors when the outter stream errors", () => {
-      scheduler().run(({ expectObservable, cold }) => {
-        const sourceStr = "--#"
-        const source = cold(sourceStr) as any
-        const result = collect<string, string>()(source)
-        expectObservable(result.get("foo")).toBe("--#")
-      })
-    })
-
-    it("completes when the outter stream completes", () => {
-      scheduler().run(({ expectObservable, cold }) => {
-        const sourceStr = "--|"
-        const source = cold(sourceStr) as any
-        const result = collect<string, string>()(source)
-        expectObservable(result.get("bar")).toBe("--|")
-      })
-    })
-  })
 })
