+++ packages/utils/src/collect.ts
@@ -1,10 +1,4 @@
-import {
-  GroupedObservable,
-  Observable,
-  OperatorFunction,
-  pipe,
-  Subscription,
-} from "rxjs"
+import { GroupedObservable, Observable, pipe } from "rxjs"
 import {
   startWith,
   endWith,
@@ -16,12 +10,6 @@ import {
 import { CollectorAction, collector } from "./internal-utils"
 
 const defaultFilter = pipe(ignoreElements(), startWith(true), endWith(false))
-
-export type CollectedObservable<K, V> = Observable<
-  Map<K, GroupedObservable<K, V>>
-> & {
-  get: (key: K) => Observable<V>
-}
 
 /**
  * A pipeable operator that collects all the GroupedObservables emitted by
@@ -35,7 +23,7 @@ export const collect = <K, V>(
   filter?: (source$: GroupedObservable<K, V>) => Observable<boolean>,
 ): ((
   source$: Observable<GroupedObservable<K, V>>,
-) => CollectedObservable<K, V>) => {
+) => Observable<Map<K, GroupedObservable<K, V>>>) => {
   const enhancer = filter
     ? (source$: GroupedObservable<K, V>) =>
         filter(source$).pipe(
@@ -45,38 +33,11 @@ export const collect = <K, V>(
         )
     : defaultFilter
 
-  const operator: OperatorFunction<
-    GroupedObservable<K, V>,
-    Map<K, GroupedObservable<K, V>>
-  > = collector((o) =>
+  return collector((o) =>
     map((x) => ({
       t: x ? (CollectorAction.Set as const) : (CollectorAction.Delete as const),
       k: o.key,
       v: o,
     }))(enhancer(o)),
   )
-
-  return (source$) => {
-    const result$ = operator(source$)
-    const get = (key: K) =>
-      new Observable<V>((observer) => {
-        let innerSub: Subscription | undefined
-        let outterSub: Subscription = result$.subscribe(
-          (n) => {
-            innerSub = innerSub || n.get(key)?.subscribe(observer)
-          },
-          (e) => {
-            observer.error(e)
-          },
-          () => {
-            observer.complete()
-          },
-        )
-        return () => {
-          innerSub && innerSub.unsubscribe()
-          outterSub.unsubscribe()
-        }
-      })
-    return Object.assign(result$, { get })
-  }
 }
