+++ packages/utils/src/index.tsx
@@ -1,5 +1,6 @@
 export { collectValues } from "./collectValues"
 export { collect } from "./collect"
+export { getGroupedObservable } from "./getGroupedObservable"
 export { createListener } from "./createListener"
 export { mergeWithKey } from "./mergeWithKey"
 export { split } from "./split"
